homework/4_LDAP_XACML/be_flask/requirements.txt
Flask>=2.0
flask-cors>=3.0
Werkzeug>=2.0
# Database
Flask-SQLAlchemy>=3.0
psycopg2-binary>=2.9
PyJWT>=2.0
# HashiCorp Vault client
hvac>=1.2.1